+++ Src/src/features/navigation/index.ts
@@ -3,7 +3,6 @@ import { MUMPS_LANGUAGE_ID } from '../../config/language';
 import { registerSymbolsFeature } from '../symbols';
 import { MumpsDefinitionProvider } from './definitionProvider';
 import { MumpsDocumentLinkProvider } from './documentLinkProvider';
-import { MumpsReferenceProvider } from './referenceProvider';
 import { MumpsNavigationHoverProvider } from './navigationHoverProvider';
 import { MumpsRoutineIndex } from './routineIndex';
 import { registerNavigationDebugCommands } from './debugCommands';
@@ -20,7 +19,6 @@ export function registerNavigationFeature(context: vscode.ExtensionContext, outp
     routineIndex,
     vscode.languages.registerDefinitionProvider(MUMPS_LANGUAGE_ID, new MumpsDefinitionProvider(routineIndex, output)),
     vscode.languages.registerDocumentLinkProvider(MUMPS_LANGUAGE_ID, new MumpsDocumentLinkProvider(routineIndex, output)),
-    vscode.languages.registerReferenceProvider(MUMPS_LANGUAGE_ID, new MumpsReferenceProvider(routineIndex)),
     vscode.languages.registerHoverProvider(MUMPS_LANGUAGE_ID, new MumpsNavigationHoverProvider(routineIndex, output))
   );
   registerNavigationDebugCommands(context, routineIndex, output);